refactor: export utiles func to packages/lib (#62)
