+++ apps/web/src/components/cta/forApp/ForApp.tsx
@@ -1,3 +1,4 @@
+import { handleSessionStorage } from '@ahhachul/lib'
 import Image from 'next/image'
 import LogoImg from 'public/illust/img/img_favicon.png'
 import { useEffect, useState } from 'react'
@@ -5,8 +6,6 @@ import { useEffect, useState } from 'react'
 import * as S from './styled'
 import { APP_CONVERSION_CTA_STORAGE_KEY } from '@/constants'
 import { useDisclosure } from '@/hooks'
-
-import handleSessionStorage from '@/utils/storage'
 
 function BottomSheetForApp() {
   const [isMounted, setIsMounted] = useState(false)
