--- apps/web/src/utils/sample.ts
@@ -1,3 +1,0 @@
-export const sampleUtils = () => {
-  return 'sample'
-}
