+++ packages/lib/src/utils/device.ts
@@ -1,0 +1,11 @@
+declare const global: unknown
+
+export const isServer = () => {
+  return typeof window === 'undefined' && typeof global !== 'undefined'
+}
+
+export function isClient() {
+  return !isServer()
+}
+
+export const isProd = (env: string): boolean => env === 'production'
