+++ packages/lib/src/utils/storage.ts
@@ -1,9 +1,7 @@
-const handleSessionStorage = (key: string) => {
+export const handleSessionStorage = (key: string) => {
   return {
     get: () => sessionStorage.getItem(key) ?? null,
     set: (newKey: string) => sessionStorage.setItem(key, newKey),
     remove: () => sessionStorage.removeItem(key),
   }
 }
-
-export default handleSessionStorage
